It works but I'd like to make this better.
Essentially, I'd like lines 66-70 to execute when the user enters a character other than 'F', 'f', 'C', 'c', 'Q' or 'q' informing them that an invalid character has been selected; thus, prompting them to give it another go. Now it works - of course when the multi-line note indicators are removed but once the user selects a correct character - per se 'f', the flow goes on to prompt them to enter the degrees Celsius, returns the corresponding degrees Fahrenheit and then still executes the else clause, stating that their entry is invalid. I assume this happening because the user enters the numeric value for - be it, degrees Celsius or degrees Fahrenheit which is obviously not one of the required characters. Moreover, I thought the flow wouldn't count the entry for the numeric value for either Celsius or Fahrenheit seeing that this execution is within the conditional if statements. Should I just resort to utilizing the exception handling technique alone? I'd like to get the else statement to work but if no one has any tricks then I'll just have to settle.